+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @memo28/utils
 
+## 1.2.1
+
+### Patch Changes
+
+- update function
+
 ## 1.2.0
 
 ### Minor Changes