+++ packages/viteBuild/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @memo28/vitebuild
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @memo28/utils@1.2.1
+
 ## 3.0.1
 
 ### Patch Changes